+++ xesmf/tests/test_frontend.py
@@ -955,7 +955,7 @@ def test_spatial_averager_mask():
 
 
 def test_densify_polys():
+    # Check that using a large poly raises a warning
     poly = Polygon([(-80, -40), (80, -40), (80, 40), (-80, 40)])  # Large poly
-    dense_poly = xe.SpatialAverager.densify_polys([poly])
-
-    assert_almost_equal(dense_poly[0].area, poly.area)
+    with pytest.warns(UserWarning):
+        xe.SpatialAverager(ds_in, [poly])